feat(settings): add advanced settings section
