+++ src/i18n/locales/cs/translation.json
@@ -10,6 +10,7 @@
   },
   "sidebar": {
     "general": "Obecné",
+    "advanced": "Advanced",
     "shortcuts": "Zkratky",
     "models": "Modely",
     "dictionary": "Slovník",
