+++ src/i18n/locales/de/translation.json
@@ -10,6 +10,7 @@
   },
   "sidebar": {
     "general": "Allgemein",
+    "advanced": "Advanced",
     "shortcuts": "Tastenkürzel",
     "models": "Modelle",
     "dictionary": "Wörterbuch",
