+++ src/i18n/locales/es/translation.json
@@ -10,6 +10,7 @@
   },
   "sidebar": {
     "general": "General",
+    "advanced": "Advanced",
     "shortcuts": "Atajos",
     "models": "Modelos",
     "dictionary": "Diccionario",
