+++ src/i18n/locales/it/translation.json
@@ -10,6 +10,7 @@
   },
   "sidebar": {
     "general": "Generale",
+    "advanced": "Advanced",
     "shortcuts": "Scorciatoie",
     "models": "Modelli",
     "dictionary": "Dizionario",
