+++ src/i18n/locales/pl/translation.json
@@ -10,6 +10,7 @@
   },
   "sidebar": {
     "general": "Ogólne",
+    "advanced": "Advanced",
     "shortcuts": "Skróty",
     "models": "Modele",
     "dictionary": "Słownik",
