+++ src/i18n/locales/pt/translation.json
@@ -10,6 +10,7 @@
   },
   "sidebar": {
     "general": "Geral",
+    "advanced": "Advanced",
     "shortcuts": "Atalhos",
     "models": "Modelos",
     "dictionary": "Dicionário",
