+++ src/i18n/locales/tr/translation.json
@@ -10,6 +10,7 @@
   },
   "sidebar": {
     "general": "Genel",
+    "advanced": "Advanced",
     "shortcuts": "Kısayollar",
     "models": "Modeller",
     "dictionary": "Sözlük",
