+++ src/i18n/locales/uk/translation.json
@@ -10,6 +10,7 @@
   },
   "sidebar": {
     "general": "Загальні",
+    "advanced": "Advanced",
     "shortcuts": "Гарячі клавіші",
     "models": "Моделі",
     "dictionary": "Dictionary",
