+++ src/i18n/locales/vi/translation.json
@@ -10,6 +10,7 @@
   },
   "sidebar": {
     "general": "Chung",
+    "advanced": "Advanced",
     "shortcuts": "Phím tắt",
     "models": "Mô hình",
     "dictionary": "Từ điển",
