+++ src/i18n/locales/zh-TW/translation.json
@@ -10,6 +10,7 @@
   },
   "sidebar": {
     "general": "一般",
+    "advanced": "Advanced",
     "shortcuts": "快捷鍵",
     "models": "模型",
     "dictionary": "字典",
